added FilterGroup tempalte class to apply multiple filter plugins to generic data structures
